Migrated the changes from CANopen demo to CANopen component
