+++ component/CANopen/CANopen.asn
@@ -32,7 +32,7 @@ CANopen-Object-Index ::= INTEGER(0..65535)
 CANopen-Subobject-Index ::= INTEGER(0..255)
 
 CANopen-Value ::= CHOICE {
-   boolean-d CANopen-Boolean,
+   boolean8 CANopen-Boolean,
    unsigned8 CANopen-Unsigned8,
    unsigned16 CANopen-Unsigned16,
    unsigned24 CANopen-Unsigned24,
@@ -73,7 +73,7 @@ CANopen-NMT-HeartbeatReason ::= ENUMERATED {
 
 Time ::= SEQUENCE {
    seconds INTEGER(0..4294967295),
-   nanoseconds INTEGER(0..4294967295)
+   nanoseconds INTEGER(0..999999999)
 }
 
 -- Helper types
@@ -98,6 +98,24 @@ Set-Data-Request ::= SEQUENCE {
 Node-Command-Request ::= SEQUENCE {
    node-id CANopen-NodeID,
    command CANopen-NMT-Command
+}
+
+CANopen-Init-Result ::= ENUMERATED {
+   success,
+   already-initialized,
+   device-creation-failure,
+   network-creation-failure,
+   nmt-creation-failure,
+   sync-creation-failure
+}
+
+CANopen-ObjDict-Operation-Result ::= ENUMERATED {
+   success,
+   device-not-initialized,
+   invalid-index,
+   invalid-subindex,
+   unsupported-type,
+   write-failed
 }
 
 END
