+++ component/CANopen/work/canopen/C/src/config.h
@@ -25,7 +25,6 @@
 /* Define to 1 if you have the <inttypes.h> header file. */
 #define HAVE_INTTYPES_H 1
 
-#ifdef CANOPEN_LINUX
 /* Define to 1 if you have the <linux/can/error.h> header file. */
 #define HAVE_LINUX_CAN_ERROR_H 1
 
@@ -40,7 +39,6 @@
 
 /* Define to 1 if you have the <linux/rtnetlink.h> header file. */
 #define HAVE_LINUX_RTNETLINK_H 1
-#endif
 
 /* Define to 1 if you have the <minix/config.h> header file. */
 /* #undef HAVE_MINIX_CONFIG_H */
@@ -51,10 +49,8 @@
 /* Define to 1 if you have the <stdint.h> header file. */
 #define HAVE_STDINT_H 1
 
-#ifdef CANOPEN_STDIO
 /* Define to 1 if you have the <stdio.h> header file. */
 #define HAVE_STDIO_H 1
-#endif
 
 /* Define to 1 if you have the <stdlib.h> header file. */
 #define HAVE_STDLIB_H 1
