+++ component/CANopen/work/canopen/C/src/debug.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#ifdef CANOPEN_DEBUG
+#include <stdio.h>
+#define DEBUG_PRINT(...) printf(__VA_ARGS__)
+#else
+#define DEBUG_PRINT(...) ((void)0)
+#endif
